requirements.txt
python-dotenv
pandas
numpy
matplotlib
scikit-learn
pymongo
certifi
pymongo[srv]==3.11
pyaml
dill